+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 - 2023-12-29
+
+### Added
+
+- **Sequenced Pulse Generator**
+- Control Chip - a crafting ingredient for the Sequenced Pulse Generator
+- Ponders for
+  - Encased Chain Cogwheel
+  - Inverted Clutch
+  - Inverted Gearshift
+  - Parallel Gearbox
+  - Sequenced Pulse Generator
+
+### Fixed
+
+- Z-fighting on the model of Encased Chain Cogwheel
+- Crash when JEI is not installed (#7)
+
 ## 0.2.0 - 2023-12-25
 
 Merry Christmas!